pilot/requirements.txt
numpy
pandas
scikit-learn
sentence-transformers
google-genai
python-dotenv
